Changes to PIN, Database Access
Fixed the Database Access bug, I think. I was just ordering commands wrong. Also removed the PIN as an identifier in the database, and replaced it conceptually as an expanded tag. There will need to be a list of master nodes somewhere, and doing that interface will be tricky, but it will simplify a lot of other things.